Added locations in Web
